test code deduplication
